RIA-9207 Refactor with mapper class for sonar scan
